Swapped compute_metric to metrics.py
